+++ src/co/edu/uptc/presenters/ManagerGeneral.java
@@ -2,7 +2,7 @@ package co.edu.uptc.presenters;
 
 import co.edu.uptc.models.Alex202128687.ManagerModelAlex;
 import co.edu.uptc.models.Model202127343.ManagerModel202127343;
-import co.edu.uptc.models.BryanModel.MyManagerModel;
+import co.edu.uptc.models.Model202022012.MyManagerModel;
 import co.edu.uptc.models.Model202127812.ManagerModel202127812;
 import co.edu.uptc.models.model202127061.ManagerModel202127061;
 import co.edu.uptc.models.ModelGerman.ManagerModel;
