+++ src/co/edu/uptc/views/people/PanelModelSelect.java
@@ -55,7 +55,6 @@ public class PanelModelSelect extends JPanel {
         configPrefersize();
         JButton jButtonSelectModelUser = new JButton("Model: student ");
         jButtonSelectModelUser.setPreferredSize(new Dimension(180,25));
-        JButton jButtonSelectModelUser = new JButton("Model: 202115100");
         add(jButtonSelectModelUser);
         jButtonSelectModelUser.addActionListener(new ActionListener() {
             @Override
